refactor(connector): extract shared wire types
Moves createReq, updateReq, configEnvelope, postgresSpec, createResp,
and getConnectorResp out of create.go / update.go into types.go.

Pure relocation — tests and coverage unchanged — but it gives the
upcoming dialect split a seam: per-dialect spec types will land as
types_snowflake.go / types_databricks.go without churning create.go
or update.go.
